+++ client/src/views/GestoreDashboardView.vue
@@ -445,7 +445,7 @@ const caricaDati = async () => {
 }
 
 const caricaGarage = async () => {
-  const res = await fetch('/api/garages-gestore', { credentials: 'include' })
+  const res = await fetch('/api/garage/garages-gestore', { credentials: 'include' })
   if (!res.ok) return
   const data = await res.json()
   mieiGarage.value = data
@@ -498,12 +498,12 @@ const aggiornaMappaOrari = async () => {
 }
 
 const caricaStorico = async () => {
-  const res = await fetch('/api/prenotazioni-gestore', { credentials: 'include' })
+  const res = await fetch('/api/prenotazioni/prenotazioni-gestore', { credentials: 'include' })
   if (res.ok) storicoPrenotazioni.value = await res.json()
 }
 
 const caricaStato = async () => {
-  const res = await fetch('/api/stato-garages-gestore', { credentials: 'include' })
+  const res = await fetch('/api/garage/stato-garages-gestore', { credentials: 'include' })
   if (res.ok) allerteStato.value = await res.json()
 }
 
@@ -519,7 +519,7 @@ const salvaNuovoGarage = async () => {
   erroreForm.value = ''
   staSalvando.value = true
   try {
-    const res = await fetch('/api/garages-gestore', {
+    const res = await fetch('/api/garage/garages-gestore', {
       method: 'POST',
       credentials: 'include',
       headers: { 'Content-Type': 'application/json' },
